Rediseño de evaluación: sliders, basePoints y preview
Se reemplaza el rating textual por sliders de puntuación base decimal para cada habilidad, con previsualización del impacto en el vector de skills. Se actualiza el modelo, comandos y tests para usar `basePoints` (`double`). El modal permite seleccionar trabajador y muestra tabla comparativa de niveles. Se mejora el flujo de transición de estados de tareas (`Finish` <-> `Archived`). Se ajustan tipos en la base de datos para mayor precisión.

+++ database/init.sql
@@ -94,10 +94,10 @@ CREATE TABLE IF NOT EXISTS performance_evaluations (
     skill_position    INT NOT NULL,
     rating            performance_rating NOT NULL,
     criticality       criticality NOT NULL,
-    base_points       NUMERIC(3,1) NOT NULL,
-    impact            NUMERIC(4,2) NOT NULL,
-    previous_level    NUMERIC(4,1) NOT NULL,
-    new_level         NUMERIC(4,1) NOT NULL,
+    base_points       DOUBLE PRECISION NOT NULL,
+    impact            DOUBLE PRECISION NOT NULL,
+    previous_level    DOUBLE PRECISION NOT NULL,
+    new_level         DOUBLE PRECISION NOT NULL,
     created_at        TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
